backend/pyproject.toml
[project]
name = "espen-api"
version = "1.0.0"
description = "ESPEN — API FastAPI para gestão por competências (SENAPPEN/DEPEN)"
requires-python = ">=3.12"
dependencies = [
  "fastapi==0.116.1",
  "uvicorn[standard]==0.35.0",
  "sqlalchemy==2.0.43",
  "psycopg[binary]==3.2.9",
  "pyjwt==2.9.0",
  "pydantic-settings==2.10.1",
  "python-multipart==0.0.20",
  "email-validator==2.2.0",
  "alembic==1.13.3",
]

[project.optional-dependencies]
dev = [
  "ruff==0.6.9",
  "pytest==8.3.3",
  "httpx==0.27.2",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.ruff]
line-length = 110
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]  # diagnóstico de comprimento já é dado pelo formatador

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
